Add redhat repository to :repositories in project.clj
Leiningen correctly resolves transitive dependencies stored in repositories
specified in the POMs/descriptors of intermediate items in the dependency tree;
however, this is not true of several other Aether-based tools for
dependency-tree interaction/recording.

Adding the "redhat" repository (needed for com.googlecode/jsendnsca-core to be
found as a dependency of cljr-nsca) thus unbreaks some supplementary/3rd-party
tooling.